Refactor header share list
